requirements.txt
gensim==4.2.0
lightning==2.0.2
pyvi==0.1.1
PyYAML==6.0
termcolor==2.3.0
torch==2.0.0
torchmetrics==0.10.0
transformers==4.28.1
underthesea==6.3.0